Support recurring events, v1
This adds support for recurring events, so long as the instance of the event
that you want to appear in the output file has been *modified*.

So, if you are using Google calendar, and you make changes, save them, and
select "only this event", it will create a special record of that event for
that particular instance.  That record will now be picked up and included in
the output file.  (It actually would before this commit if the modification
changed the duration.  Now it will be included in the output of there is any
modification at all, cause it now knows how to get the duration from the
initial "leader" event (the one with the RRULE).)